feat: replace unpack implementation to be multithreaded by default
