+++ src/models/commands/mod.rs
@@ -1,7 +1,5 @@
 mod list_command;
-mod unpack_command;
-mod unpack_async_command;
+mod unpack;
 
 pub use list_command::*;
-pub use unpack_command::*;
-pub use unpack_async_command::*;
+pub use unpack::*;
