+++ src/traits/mod.rs
@@ -1,5 +1,7 @@
 mod parse_utf8;
 mod pure_runnable;
+mod seek_read;
 
 pub use parse_utf8::*;
 pub use pure_runnable::*;
+pub use seek_read::*;
