+++ src/traits/pure_runnable.rs
@@ -1,3 +1,7 @@
 pub trait PureRunnable {
 	fn run(&self);
 }
+
+pub trait PureRunnableAsync {
+	fn run(&self) -> impl std::future::Future<Output = ()> + Send;
+}
